[eradius] generation of nas id, get back documentation and tests
Summary:
There is a NasId, that was written before, but can be now generated from generic name + ip. This allows to eliminate nas id per name.
As backwards compatibility it is still possible to specify a nas_id. Documentation is added back. Tests, that test configuration are added.

Test Plan: tetrapak test

Reviewers: vtarasen, orixmann, aschultz

CC: mkraemer, aschultz

Differential Revision: https://review.tpip.net/D1370