use prometheus_histogram:declare/1 instead prometheus_histogram:new/1
When a RADIUS request comes to eradius or we try to send a request
from it - new histogram metric could be created. There could be
a situation when a couple of RADIUS requests with the same name
and similar labels come to eradius which may lead to crash because
of races during creation of prometheus metric.

This commit replaces prometheus_histogram:new/1 with
prometheus_histogram:declare/1 API as the first one throws an exception
in a case of metric already exists.