Fix gathering of RADIUS Accounting metrics
We are using ets:match_spec_run/2 to go over given attributes to
match type of RADIUS Accounting request to update a certain metric by
RADIUS Accounting type.

The issue is that we may get RADIUS attributes during sending of a RADIUS
request in two following formats:

   * [{AttrId, Value}, ...]
   * [#attribute{}, ....]

But only the first one could be matched in ets:match_spec_run/2. Thus if
RADIUS attributes will be given in the second format - a RADIUS Accounting
client metric will not be updated.

This commit fixes this situation.