Improve RADIUS fail-over mechanism
When RADIUS client process is started it collects all secondary RADIUS
servers from RADIUS server pools into ets table to use them for fail-over.

This commit also adds primary servers to this ets table. We need this
to avoid following situation:

A primary RADIUS server was used but eradius client failed to send requests
there by some reasons. So such RADIUS server will be marked as inactive for
a some configurable period of time and secondary RADIUS server will be used.
The issue is that eradius client could be used again with the primary server
while it is still in inactive state.

This commit adds all primary and secondary servers to an ets table that stores
endpoints of RADIUS servers that could be used. So if eradius client will be
called with currently inactive RADIUS server - it will not be used if there is
another possible active RADIUS server exists.

Backward compatibility is preserved and new functionality will be used only
in a case when fail-over is enabled.