Make server status metrics optional
This commits adds new configuration option - `server_status_metrics_enabled` to
enable new RADIUS server status metrics.

This was done to avoid situation when eradius is used without prometheus library
so creation/usage of this metric will not be crashed at startup of eradius
application.

The default value of the new configuration option is set to `false` to preserve
backward compatibility.